+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on and uses the types of changes according to [Keep a Change
 - Added a new private function `Measure-AuditResult` to calculate and display audit results.
 - Enhanced error logging to capture failed test details and display them at the end of the audit.
 - Added a private function `Get-RequiredModule` to initialize the `$requiredModules` variable for better code organization in the main script.
-- Updated `Test-MailboxAuditingE3` and `Test-MailboxAuditingE5` functions to use `Format-MissingActions` for structuring missing actions into a pipe-separated table format.
+- Updated `Test-MailboxAuditingE3` and `Test-MailboxAuditingE5` functions to use `Format-MissingAction` for structuring missing actions into a pipe-separated table format.
 - Added more verbose logging to `Test-BlockMailForwarding` and improved error handling for better troubleshooting.
 - Improved `Test-RestrictCustomScripts` to handle long URL lengths better by extracting and replacing common hostnames, and provided detailed output.
 - Added sorting to output.